fix(cli): share positive-int parser and apply it to template create
Moves parsePositiveInt into options.ts and uses it for create's
--cpu-count/--memory-mb, which previously used bare parseInt and
silently accepted NaN.
