+++ packages/cli/src/options.ts
@@ -1,6 +1,22 @@
 import * as commander from 'commander'
 
-import { asBold } from './utils/format'
+import { asBold, asLocal } from './utils/format'
+
+/**
+ * Parse a CLI option as a positive integer, rejecting non-numeric values so
+ * they don't silently become NaN.
+ */
+export function parsePositiveInt(label: string): (value: string) => number {
+  return (value) => {
+    const parsed = Number(value)
+    if (!Number.isInteger(parsed) || parsed < 1) {
+      throw new commander.InvalidArgumentError(
+        `${label} must be a positive integer. You provided ${asLocal(value)}.`
+      )
+    }
+    return parsed
+  }
+}
 
 export const pathOption = new commander.Option(
   '-p, --path <path>',
